refactor(minecat): move non-cogs out of cog
