--- minecat/cogs/websocket/__init__.py
@@ -1,10 +1,0 @@
-# Copyright 2022-latest Oliver Wilkes. Licensed under the Elastic License 2.0;
-# you may not use this file except in compliance with, the Elastic License 2.0
-# https://www.elastic.co/licensing/elastic-license
-
-from __future__ import annotations
-
-from ._errors import *
-from ._manager import *
-from ._opcode import *
-from ._server import *
