+++ minecat/websocket/__init__.py
@@ -1,3 +1,10 @@
 # Copyright 2022-latest Oliver Wilkes. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with, the Elastic License 2.0
 # https://www.elastic.co/licensing/elastic-license
+
+from __future__ import annotations
+
+from .errors import *
+from .manager import *
+from .opcode import *
+from .server import *
