+++ minecat/websocket/errors.py
