+++ minecat/websocket/opcode.py
